+++ src/logging/winston/log-level.ts
@@ -1,0 +1,8 @@
+export enum LogLevel {
+  error = 'error',
+  warn = 'warn',
+  info = 'info',
+  verbose = 'verbose',
+  debug = 'debug',
+  silent = 'silent',
+}
